Remove domain because concept is not cohesive enough here. Split tree mutations into actions and operations. Cleanup dependencies.
