+++ src/utils/createId.ts
@@ -1,0 +1,3 @@
+export const createId = (): string => {
+  return Math.random() + "";
+};
